feat: #18 사진 리스트 생성
