+++ src/assets/styles/Photo.css
@@ -1,0 +1,14 @@
+#photo-wrapper {
+  overflow: scroll;
+}
+
+#photo-list {
+  display: flex;
+}
+
+img {
+  width: 60px;
+  height: 60px;
+  padding-left: 10px;
+  padding-right: 10px;
+}
