+++ src/utils/getPhotos.ts
@@ -1,0 +1,13 @@
+import img1 from '../assets/images/img1.jpg';
+import img2 from '../assets/images/img2.jpeg';
+import img3 from '../assets/images/img3.png';
+import img4 from '../assets/images/img4.jpeg';
+import img5 from '../assets/images/img5.jpeg';
+import img6 from '../assets/images/img6.jpeg';
+import img7 from '../assets/images/img7.jpeg';
+import img8 from '../assets/images/img8.jpeg';
+import img9 from '../assets/images/img9.jpeg';
+import img10 from '../assets/images/img10.jpeg';
+
+const photos = { img1, img2, img3, img4, img5, img6, img7, img8, img9, img10 };
+export default photos;
