+++ src/views/Alarm.ts
@@ -1,15 +1,6 @@
 import Component from '../core/Component';
 import '../assets/styles/Alarm.css';
-import { InlineFunctions } from 'uglify-js';
-
-const getSequenceArray = (start, end, interval) => {
-  const temp = [];
-  for (let i = start; i < end; ) {
-    temp.push(('0' + i.toString()).slice(-2));
-    i += interval;
-  }
-  return temp;
-};
+import { getSequenceArray } from '../utils/getSequenceArray';
 
 export default class Alarm extends Component {
   setup() {
